feat: exception raised when unknown release returned for release flag
I think this is the right approach for now. It's possible that it might be good to configure this behaviour. I can image use cases where people want true or value as a default or to run some conditional check on whether to raise an exception (e.g. !Rails.env.production?).